refactor: extract useSetI18nConfig hook
